Fixes for multicore
- Don't de-initialize the OCOTP after checking fuses -- MCMGR relies on
it being initialized
- Adjust IPC to not block while waiting for the remote core -- this lets
the task sleep and other things to progress if anything goes wrong on
the M4.
- Clean up the linker memory allocations for both cores.
- Re-add core1 sections to the m7 binary -- note that you probably can't
flash this onto the device, only RAM load. The core1 image lives in a
much higher memory space, and the tools try and create empty space
between and generate huge binaries. Long-term, it will probably be the
case that we want to reserve blocks on the NAND for the M4 image, and
load it from NAND->RAM explicitly.
- Don't generate the sbfile in the flashtool if we're loading to ram -
not necessary and saves a lot of time if you have a core1 image ;)

GitOrigin-RevId: 5f79500b184f7cc6b9ec78de14c0597b111261ab