ELF Bootloader
- Implement a bootloader application that can read an application in ELF
format, and execute it.
  - Can either read from the flash storage of the device or USB HID.
  - If no HID transfer is started, will automatically boot default.elf
  from the filesystem.
- Refactor the NXP MCMGR initialization into a library, so ELFLoader can
exclude it -- that code doesn't handle being re-executed nicely, and
ELFLoader isn't going to use both cores anyways.
- Allow M7 applications to specify a custom linker script (as ELFLoader
does). If none is specified, the default linker script is used.
- Drop the HelloWorldMultiCore bare-metal sample.
- Make USB descriptors "dynamic" -- they need to be registered before we
initialize the USB stack, but applications are no longer all forced to
be stuck with the same descriptors if they're not using a feature. This
allows ELFLoader to only expose HID, and other apps to still expose CDC
ACM/EEM.
- Update flashtool to handle the HID transfer, and adjust the filesystem
flashing path to use ELFLoader as the ROM booted app, and the
user-supplied app as default.elf.
- Update udev rules to give permissions for HID.

GitOrigin-RevId: 4787c6d0cd1436a1e38cf9b49a816baa503492c9