Beginnings of OOBE demo, misc multicore patches
- Add the starts of our OOBE demo. This runs a person detection model on
the M4, which wakes the M7 if it detects anyone. The M7 then executes
Posenet w/ the TPU. If the posenet model does not detect any
sufficiently good poses for 5 seconds, the M7 stops executing the model,
and the M4 runs person detection again.
- Adjustments to the Enable/Disable of the camera driver, to allow
passing it between cores. At the moment, this requires the core using
the camera to stop before the other core tries to take over, smarter
handoff may come later.
- Shrink the M7 heap space and allocate that space to the M4's heap.
- Use some of the shared OCRAM for M4's BSS section
- Clean up a few CMake targets to compile separately for M4 and M7
- Instantiate filesystem on M4.

Change-Id: I3d6f341eb8a2a8311938757ccaac9df8ad87e154
GitOrigin-RevId: a9d5bfea2785b0a8deff451704de73018e5c833b