Add MFGTestWiFi for WiFi/BT
- Add MFGTest for the WiFi/BT board. Implements wifi_get_ap to scan for
access points, wifi_set_antenna to switch between internal and external
antenna, and ble_scan to scan for a specified bluetooth MAC.
- Due to some oddness with the bluetooth stack and consoles, there are a
few tweaks to main to allow disabling the console input and output.
- Move the define for setting the debug console to non-blocking into a
configurable CMake variable, so that we can switch it on and off
(interacts badly with Bluetooth, as well).
- Add a new flag to flashtool, --subapp. This allows selecting an
application that does not match the name of the directory in apps/

Change-Id: Ibdc47354704ca3bd81f35d6766ec0a833af70119
GitOrigin-RevId: 55ecfaf47dc13b345823e32e1138d1c7d56081be