Add flag to use cached file list in flashtool
Flashtool uses os.walk to recursively search for all the libraries
needed by a binary when creating the filesystem. Add a new flag,
--cached_files_path which points to a text file that contains all all
the files in the build directory.

For now, this file is created using os.walk and outputting to a file,
but it should eventually be added to the build.

Tested:
Flashtool with and without the cached file list

Change-Id: I01b61210385d59a2d4eda7a8aac6b1f9f5a7d4d6
GitOrigin-RevId: a550e3b389d973558bfec0cdc0dd25c94a6a807a