Have ELFLoader halt loading if the ELF is missing
- It's possible to have ELFLoader on a device, but no application to run
-- whether through a flashing bug, deleting the ELF, etc.
- ELFLoader never checks that the ELF exists -- it assumes that it
either loaded correctly from the filesystem or via USB. Add checks to
see if there is a valid ELF of greater than zero size. If it's not
there, suspend the task which loads and jumps into the ELF.
- Slightly adjust when we disable USB -- don't do it until we know we
have an ELF. This way, ELFLoader can still safely accept a new payload.

Change-Id: I88b85e988fe17b478e41bacf42555a5653ebc59e
GitOrigin-RevId: f75c0412d7779e04e1431359eaa6f5b7561f1159