Minor CMake + build script fixups to add ninja support
- Fixups to let us build w/ Ninja instead of Make: primarily intended
to let us easily build on Windows, but no reason other platforms couldn't
use Ninja as well.
- Drop the find_program for make: CMake does this already,
and in a way that supports different backend tools properly.
- Drop an add_custom_command from the M4 executable rule:
this was generating duplicate rules for one target. Make doesn't
seem to care about that, but Ninja does. No impact on the output, M4
binaries are still built and included correctly -- so this line probably
was never necessary.
- Add flags to the build scripts, and install ninja in the Dockerfile so we can test.

Change-Id: I66d3bbda827b6bc5ccd7d8490031235dde0de290
GitOrigin-RevId: be4f900d25b6971ce97a1940b53c15b883b29664