Windows flashing fixes
- Loosen up an assert in ELFLoader: HID packets may be larger than
what we expect to be contained in them, and all but one of our asserts
accounted for that. Fix the straggler that was expecting exact length match.
- Only check for baud rate on the CdcAcm reset: Windows doesn't necessarily
configure the port in the same way as Linux, wrt DTE and carrier: so don't
check them. This makes the port reset faster as well, so flashtool must
catch an exception for the serial port going away now (legacy behavior is
preserved, in case edge cases may need it).
- Libusb on Windows has some issues getting strings for composite devices:
so it can't find the device when an application is running. Instead of checking
the USB descriptor strings for the serial in app mode, check the serial port for
its serial number -- which is the same as the USB serial!

Change-Id: Idab3e6edd762cf2e548a03a48365743ea9420ed8
GitOrigin-RevId: 3a324a2224ee43d5b4ada0e28b14950e09910c12