Add OOBE Demos
Adds a demo version of OOBE for use when there aren't going to be faces
in front of the camera that trigger the M4->M7 transition. This is good
for demoing as well as regression testing.

Also includes a OOBEDemoEthernet and OOBEDemoWifi which currently just
connect to the network but don't do anything else. Eventually, the OOBE
output should be available on these networks and it may be worth pulling
into the main OOBE.

Change-Id: Id1a5070700aba78dad2daa9c673cd2b621e7c167
GitOrigin-RevId: 1397a9e431e35b839ce5387f9e53d1e6b487a83f