Fix OOBE to use Camera data for PoseNet
Previously, QueueFrame was called on a copy of input when signaling to
PosenetTask. However, input was just cleared with a std::move to
camera_output. This means it was continually running on the canned test
data. Make the frame queued a copy of camera_output (and rename
camera_output in PosenetTask to avoid confusion with the global one).

Change-Id: I48e7fd033d0a5eb86d5781abe0d47e1c15e22f1c
GitOrigin-RevId: a7303125f87d6f08685e286bf48f0562fb00aa00