Allows destination to cmake's DATA rules during flash.
Some of our libraries requires data files such as firmware that are
flashed to the board by the DATA rule. This is fine for any app within
the project, however from outside of the project, the data path is
different on the device. This change allows users to define the absolute
destination of the needed file. If destination is not defined, then
destination = source.

Change-Id: I5e51d3bccddf47ff3c965f0edb04148158c7d488
GitOrigin-RevId: 9909a56e91a7ee4789876ad26f82bf7ef2f6f49f