Refactor Posenet to be adapter.
See b/235529183, b/234636681
This change remove lib_posenet and move all of the posenet helpers to
lib/tensorflow along with classification and detection. Reasons:
- Users can now switch and choose model to use without making changes
  to the library
- Users have access to all APIs provided by tf::micro and not just our
  setup & loop when using posenet models
- Users are no longer forced to run the startup invoke on test input

Unforunately, doing this does means that users of the library will
have to do their own set up of the intepreter. This is how
classification and detection works. This change affected the RackTest,
OOBE demo, and the Posenet example.

Change-Id: I29996fe412502b418970183d6f51260e24c50cd1
GitOrigin-RevId: cef4c89947628e56ebe5f9ae5e4e82129832b08c