remove redundant args and set default build_dir
NOTE: Because CMake defines the 'build' dir, if we want to
specify a default build_dir in flashtool, then it's appropriate
that we create a CMakePreset to define this path. In this manner,
instead of specifying the build path for cmake and then also
specify it everytime with flashtool, users instead use this preset
and then don't need to specify build_dir with flashtool anymore.

That is, instead of:
  cmake -B build -S .
Users use:
  cmake -S . --preset default
Just to ensure "build" is exactly as expected by flashtool.

bug: 239056228
bug: 239056536

Change-Id: I136093841110f0d0daa6fe07cf0214bfdd1c2c41
GitOrigin-RevId: 919f5e469e01b3e0f2b51db6115ebd7f9b8ee97f