Fix bug in YAMNet
After a refactor, YAMNet would load both audio data and spectrogram data
into the same input tensor. This mostly worked because the audio sample
capture and the spectrogram generation occur sequentially. But, the
input tensor is sized for the spectrogram which means the much larger
audio data was just writing memory beyond that point. It was luck that
it didn't write over anything important.

Adds an array for the audio data.

Tested:
Yamnet

Change-Id: Iec9c143fa287ce65f30e4ab11776226e0d370e09
GitOrigin-RevId: fc8d27a1d585cf8b5fd2a1d7f13d5e1b0e32d31b